+++ src/api-routes.js
@@ -1,6 +1,6 @@
 import { userApi } from "./api/user-api.js";
 import { collectionApi } from "./api/collection-api.js";
-import { trackApi } from "./api/track-api.js";
+import { placeApi } from "./api/place-api.js";
 
 export const apiRoutes = [
   { method: "GET", path: "/api/users", config: userApi.find },
@@ -16,9 +16,9 @@ export const apiRoutes = [
   { method: "GET", path: "/api/collections/{id}", config: collectionApi.findOne },
   { method: "DELETE", path: "/api/collections/{id}", config: collectionApi.deleteOne },
 
-  { method: "GET", path: "/api/tracks", config: trackApi.find },
-  { method: "GET", path: "/api/tracks/{id}", config: trackApi.findOne },
-  { method: "POST", path: "/api/collections/{id}/tracks", config: trackApi.create },
-  { method: "DELETE", path: "/api/tracks", config: trackApi.deleteAll },
-  { method: "DELETE", path: "/api/tracks/{id}", config: trackApi.deleteOne },
+  { method: "GET", path: "/api/places", config: placeApi.find },
+  { method: "GET", path: "/api/places/{id}", config: placeApi.findOne },
+  { method: "POST", path: "/api/collections/{id}/places", config: placeApi.create },
+  { method: "DELETE", path: "/api/places", config: placeApi.deleteAll },
+  { method: "DELETE", path: "/api/places/{id}", config: placeApi.deleteOne },
 ];