+++ src/models/joi-schemas.js
@@ -21,27 +21,27 @@ export const UserSpecPlus = UserSpec.keys({
 
 export const UserArray = Joi.array().items(UserSpecPlus).label("UserArray");
 
-export const TrackSpec = Joi.object()
+export const PlaceSpec = Joi.object()
   .keys({
     title: Joi.string().required().example("Piano Sonata No. 7"),
     artist: Joi.string().required().example("Beethoven"),
     duration: Joi.number().allow("").optional().example(12),
     collectionid: IdSpec,
   })
-  .label("Track");
+  .label("Place");
 
-export const TrackSpecPlus = TrackSpec.keys({
+export const PlaceSpecPlus = PlaceSpec.keys({
   _id: IdSpec,
   __v: Joi.number(),
-}).label("TrackPlus");
+}).label("PlacePlus");
 
-export const TrackArraySpec = Joi.array().items(TrackSpecPlus).label("TrackArray");
+export const PlaceArraySpec = Joi.array().items(PlaceSpecPlus).label("PlaceArray");
 
 export const CollectionSpec = Joi.object()
   .keys({
     title: Joi.string().required().example("Beethoven Sonatas"),
     userid: IdSpec,
-    tracks: TrackArraySpec,
+    places: PlaceArraySpec,
   })
   .label("Collection");